fix: Modify request behavior
Current behavior: New duplicate request to the REST API with a custom message on each browser or editor page reload.
Expected behavior: Single request to the REST API on the installation of the custom block.
Fix: Added an activation hook for the rest_client to execute only once on installation.